Make the network allocator optional.
This allows scenarios where this is not needed (i.e. non-CNM base network
configurations) to opt out and avoid the possibility of looking for CNM state
in non-CNM network specs and objects.

This wasn't as simple as just removing the network allocator since nothing else
handles the voting to propagate a task from NEW to PENDING. So this commit adds
a "backstop" allocator which does not allocate anything but just votes and
cranks the state if appropriate.

The new "backstop" allocator always runs. In the case where the network
allocator also runs it is the case (as it always was) that all voters must vote
and that the last one to do so will take care of cranking the state.

I moved allocatedStatusMessage and updateTaskStatus from network.go to
allocator.go since they are generic.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>