WIP: Add a special allocator to create network attachments
The executor needs this to happen on the managers (since executors are on the
worker/agent side they cannot see the store), it was done in the main network
allocator, but I just disabled that, it also does more than we need.

Eventually it might be interesting to see if this could be factored out of the
main allocator, not clear if this is possible.

The simple allocate seems to see UpdateTask as well as CreateTask.

Some of the handling of state transitions here is a bit subtle. I'm unclear if
I got it even half right.

Also, shouldn't share watches between allocators it seems. Not sure if backstop
needs to handle other events too, e.g. UpdateTask. Not sure if all of them
don't need to handle DeleteTask too.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>